lib: export ParseDockerURL
To allow that, move docker_functions.go code to a "docker" package and
make the exported ParseDockerURL call the internal ParseDockerURL in the
"docker" package.
